requirements-cloud.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
anthropic>=0.30.0
psycopg2-binary==2.9.9
cloud-sql-python-connector[pg8000]==1.12.1
pg8000==1.31.2
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
playwright>=1.40.0
flask-login>=0.6.3
